Move program components example to root level examples folder
