+++ examples/component-program/PulumiPlugin.yaml
