+++ examples/component-program/consumer/Pulumi.yaml
