+++ examples/component-program/nested/nestedRandomComponent.go
