+++ examples/component-program/randomComponent.go
